backend/requirements.txt
fastapi
uvicorn
pytubefix
google-generativeai
python-multipart
requests
python-dotenv
